Create main.rb for entry points
